install: Fix "Arithmetic overflow: Using operator '+' on a 4 byte value and then casting the result to a 8 byte value."
Fix "Arithmetic overflow: Using operator '+' on a 4 byte value and then
casting the result to a 8 byte value."

+++ install/nfs_install.c
@@ -101,7 +101,7 @@ int main(int ac, char *av[])
     DPRINTF(1, ("# RegQueryValueExA(), result=%d, dataSize=%d\n",
         (int)result, (int)dataSize));
 
-    originalValue = malloc(dataSize+1);
+    originalValue = malloc((size_t)dataSize+1);
 
     if (!originalValue) {
         (void)fprintf(stderr, "%s: Out of memory\n", av[0]);
